pyproject.toml
[project]
name = "hvzbot"
version = "0.1.0"
description = "A bot to report the feed of tags of RITHvZ"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "discord-py>=2.6.3",
    "lxml>=6.0.1",
    "pandas>=2.3.2",
]
